tests: Add Positive vertex_input.cpp
